remount items controller
