+++ config/application.rb
@@ -89,7 +89,6 @@ class UsefulMusic::App
   controller '/sessions', SessionsController
   controller '/password_resets', PasswordResetsController
   controller '/pieces', PiecesController
-  controller '/items', ItemsController
   controller '/purchases', PurchasesController
   controller '/shopping_baskets', ShoppingBasketsController
   controller '/orders', OrdersController
